group delete stop if it has users

+++ app/Http/Datatables/GroupDatatable.php
@@ -58,9 +58,9 @@ class GroupDatatable extends Datatable {
             if($this->checkMenuPermission($this->nav, 'permission')) {
                 $permission = "<a class='btn btn-sm btn-outline yellow-casablanca' title='Permission' href='" . route('group_permission', [$result->id]) . "'> <i class='icon-key'></i> Permission</a>";
             }
-            // if($this->checkMenuPermission($this->nav, 'delete')) {
-            //     $delete = "<a class='btn btn-sm btn-outline red table-row-delete' title='Remove' href='javascript:;' data-id='" . $result->id . "'> <i class='icon-trash'></i> Remove</a>";
-            // }
+            if($this->checkMenuPermission($this->nav, 'delete')) {
+                $delete = "<a class='btn btn-sm red table-row-delete' title='Remove' href='javascript:;' data-id='" . $result->id . "'> <i class='icon-trash'></i> Remove</a>";
+            }
             $data[] = [
                 $i+1,
                 isset($result->name) ? "<a class='' title='View' href='" . route('users_group', [$result->id]) . "'> ".$result->name."</a>" : "",
@@ -80,9 +80,13 @@ class GroupDatatable extends Datatable {
         if (isset($request->actionType) && $request->actionType == "delete_action") {
             $group = Group::find($request->record_id);
             if ($group) {
-                $group->delete();
-                $this->status = "OK";
-                $this->message = "Group deleted successfully";
+                if($group->users->count() == 0) {
+                    $group->delete();
+                    $this->status = "OK";
+                    $this->message = "Group deleted successfully";
+                }else {
+                    $this->message = "Group delete failed. Users are assigned to this group.";
+                }
             } else {
                 $this->message = "Group delete failed";
             }
